Use an "output map" for the `generated_inputs` output group
With this change a file that lists the locations of all of the generated input files is included in the `generated_inputs` output group instead of the files themselves. This results in two improvements:

1. The bazel output is cleaner. With lots of generated inputs the output from `bazel build` could be very noisy.
2. The more important thing: only the output map is listed in the BEP as an "important output". This keeps the size of the BEP small (sometimes on the order of hundreds or more megabytes smaller).

Since the files themselves are no longer listed as top-level outputs, we have to do some "trickery" to ensure they are still downloaded when using `--remote_download_toplevel`. We write a unique value to a file inside of project's internal directory, and use that file as an input to the output map generation. This, along with specific execution infos and input declarations, cause the files to be downloaded to the host as part of writing the output map.